% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackplus.R
\name{conformal.multidim.jackplus}
\alias{conformal.multidim.jackplus}
\title{Multivariate Response Jackknife + Prediction Regions}
\usage{
conformal.multidim.jackplus(x, y, x0, train.fun, predict.fun, alpha = 0.1)
}
\arguments{
\item{x}{The feature variables, a matrix n x p.}

\item{y}{The matrix of multivariate responses (dimension n x q)}

\item{x0}{The new points to evaluate, a matrix of dimension n0 x p.}

\item{train.fun}{A function to perform model training, i.e., to produce an
estimator of E(Y|X), the conditional expectation of the response variable
Y given features X. Its input arguments should be x: matrix of features,
and y: matrix of responses.}

\item{predict.fun}{A function to perform prediction for the (mean of the)
responses at new feature values. Its input arguments should be out: output
produced by train.fun, and newx: feature values at which we want to make
predictions.}

\item{alpha}{Miscoverage level. Default for alpha is 0.1.}
}
\value{
A list with components x0, lo and up. lo and up are matrices of
  dimension n0 x q containing the lower and upper bounds of the prediction
  region for each test point.
}
\description{
Compute prediction regions using multivariate Jackknife + inference.
}
\details{
The work is an extension of the univariate approach to jackknife +
 inference to a multivariate context, exploiting the concept of depth measures.
 For each test point x0, the n leave-one-out models give the 2n vectors
 mu_{-i}(x0) -/+ R_i, where R_i is the vector of absolute leave-one-out
 residuals of observation i. The vectors are standardised component-wise and
 ranked according to the depth -max_j |z_j|; the ceiling(2n(1-alpha/2))
 deepest vectors are retained and their axis-aligned bounding box is
 returned. The finite sample guarantee of the univariate jackknife+ is not
 established for this extension, and the coverage of the region can be
 below 1-alpha, in particular when q is large relative to n.

This function is based on the package future.apply to
 perform parallelisation.
}
\examples{
n=20
p=2
q=2
mu=rep(0,p)
x = mvtnorm::rmvnorm(n, mu)
beta<-sapply(1:q, function(k) c(mvtnorm::rmvnorm(1,mu)))
y = x\%*\%beta + t(mvtnorm::rmvnorm(q,1:n))
x0=x[n,,drop=FALSE]
funs=lm_multi()

sol<-conformal.multidim.jackplus(x[-n,],y[-n,],x0,train.fun = funs$train.fun,
                                 predict.fun = funs$predict.fun)
}
\references{
Barber, Candes, Ramdas, Tibshirani (2021), "Predictive inference
  with the jackknife+", The Annals of Statistics 49(1), 486-507.
}
